feat(#473): remove authors preview for comments
